requirements.txt
aniso8601==9.0.1
attrs==21.4.0
blis==0.7.7
cached-property==1.5.2
catalogue==2.0.7
certifi==2022.6.15
charset-normalizer==2.0.12
clams-python==0.5.1
click==7.1.2
cymem==2.0.6
deepdiff==5.2.3
Flask==1.1.4
Flask-RESTful==0.3.8
gunicorn==20.1.0
idna==3.3
isodate==0.6.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
lapps==0.0.2
lxml==4.9.0
MarkupSafe==2.0.1
mmif-python==0.4.6
murmurhash==1.0.7
numpy==1.22.4
ordered-set==4.0.2
packaging==21.3
pathy==0.6.1
platformdirs==2.5.2
preshed==3.0.6
pydantic==1.8.2
pyparsing==3.0.9
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2022.1
requests==2.28.0
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.16.0
smart-open==5.2.1
spacy==3.1.2
spacy-legacy==3.0.9
srsly==2.4.3
thinc==8.0.17
tqdm==4.64.0
typer==0.3.2
typing-extensions==4.2.0
urllib3==1.26.9
wasabi==0.9.1
Werkzeug==1.0.1
zeep==4.1.0
zipp==3.1.0
